Add tests for PropertiesBasedJwtSignerRegistry
