+++ src/main/java/com/contentgrid/gateway/security/jwt/issuer/JwtInternalIssuerConfiguration.java
@@ -3,16 +3,16 @@ package com.contentgrid.gateway.security.jwt.issuer;
 import com.contentgrid.gateway.security.jwt.issuer.JwtInternalIssuerConfiguration.ContentgridGatewayJwtProperties;
 import com.contentgrid.gateway.security.jwt.issuer.actuate.JWKSetEndpoint;
 import com.nimbusds.jose.JWSAlgorithm;
-import com.nimbusds.jose.jwk.JWKSet;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotNull;
 import java.util.HashMap;
 import java.util.Map;
-import java.util.Optional;
 import java.util.Set;
-import java.util.concurrent.ConcurrentHashMap;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Data;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.boot.actuate.autoconfigure.endpoint.condition.ConditionalOnAvailableEndpoint;
@@ -22,7 +22,6 @@ import org.springframework.cloud.gateway.config.conditional.ConditionalOnEnabled
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.core.io.support.ResourcePatternResolver;
 
 @RequiredArgsConstructor
 @Configuration(proxyBeanMethods = false)
@@ -47,49 +46,6 @@ public class JwtInternalIssuerConfiguration {
         return new LocallyIssuedJwtGatewayFilterFactory(applicationContext, jwtSignerRegistry::getRequiredSigner);
     }
 
-    @RequiredArgsConstructor
-    static class PropertiesBasedJwtSignerRegistry implements JwtSignerRegistry {
-        private final ContentgridGatewayJwtProperties properties;
-        private final ResourcePatternResolver resourcePatternResolver;
-        private final Map<String, JwtClaimsSigner> instantiatedSigners = new ConcurrentHashMap<>();
-
-        @Override
-        public boolean hasSigner(String signerName) {
-            return properties.getSigners().containsKey(signerName);
-        }
-
-        @Override
-        public Optional<JWKSet> getJWKSet(String signerName) {
-            if(hasSigner(signerName)) {
-                return Optional.ofNullable(getRequiredSigner(signerName).getSigningKeys().toPublicJWKSet());
-            }
-            return Optional.empty();
-        }
-
-        @Override
-        public Optional<JwtClaimsSigner> getSigner(String signerName) {
-            if(!hasSigner(signerName)) {
-                return Optional.empty();
-            } else {
-                return Optional.of(getRequiredSigner(signerName));
-            }
-        }
-
-        @Override
-        public JwtClaimsSigner getRequiredSigner(String signerName) {
-            return instantiatedSigners.computeIfAbsent(signerName, this::createSigner);
-        }
-
-        private JwtClaimsSigner createSigner(String signerName) {
-            var signerProperties = properties.getSigners().get(signerName);
-            if(signerProperties == null) {
-                throw new IllegalArgumentException("No JWT signer named '%s'. Available signers are %s".formatted(signerName, properties.getSigners().keySet()));
-            }
-            return new PropertiesBasedJwtClaimsSigner(signerProperties, resourcePatternResolver);
-        }
-
-    }
-
     @ConfigurationProperties("contentgrid.gateway.jwt")
     @Valid
     @Data
@@ -101,10 +57,18 @@ public class JwtInternalIssuerConfiguration {
 
 
     @Data
+    @Builder
+    @AllArgsConstructor
     static class JwtSignerProperties implements PropertiesBasedJwtClaimsSigner.JwtClaimsSignerProperties {
         @NotNull
         private String activeKeys;
         private String allKeys;
+        @Builder.Default
         private Set<JWSAlgorithm> algorithms = Set.of(JWSAlgorithm.RS256);
+
+        public JwtSignerProperties() {
+            this.algorithms = Set.of(JWSAlgorithm.RS256);
+        }
+
     }
 }
